\docType{data}
\name{pizzasizes}
\alias{pizzasizes}
\title{pizzasizes
A data.frame with the data for
http://www.npr.org/blogs/money/2014/02/26/282132576/74-476-reasons-you-should-always-get-the-bigger-pizza}
\description{
pizzasizes A data.frame with the data for
http://www.npr.org/blogs/money/2014/02/26/282132576/74-476-reasons-you-should-always-get-the-bigger-pizza
}
\keyword{datasets}

